backend/requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.2
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3